requirements.txt
click
numpy
scipy
python-dotenv
PyMuPDF
sentence-transformers
pdfplumber
pytest
faiss-cpu
langchain_community
langchain_text_splitters
